requirements.txt
tornado~=6.1
PyYAML~=5.3.1
kazoo~=2.8.0
ruamel.yaml
redis~=3.5.3
setuptools